Added support for dynamic vite config switching
