+++ vite.config.ts
@@ -1,53 +1,40 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import path from "path";
 
 const host = process.env.TAURI_DEV_HOST;
-const linkedPackagePath = path.resolve(__dirname, "../brc100-ui-react-components/src");
 
-export default defineConfig({
+// https://vitejs.dev/config/
+export default defineConfig(async () => ({
   plugins: [
     react({
-      include: [
-        "**/*.{jsx,tsx,js,ts}",                                // your app
-        `${linkedPackagePath}/**/*.{jsx,tsx,js,ts}`            // your linked package
-      ],
+      // Ensure that all .jsx, .tsx, .js, and .ts files are included.
+      include: "**/*.{jsx,tsx,js,ts}",
     }),
   ],
 
+  // Vite options tailored for Tauri development and only applied in `tauri dev` or `tauri build`
+  //
+  // 1. prevent vite from obscuring rust errors
   clearScreen: false,
-
+  // 2. tauri expects a fixed port, fail if that port is not available
   server: {
     port: 1420,
     strictPort: true,
     host: host || false,
     hmr: host
-      ? { protocol: "ws", host, port: 1421 }
+      ? {
+        protocol: "ws",
+        host,
+        port: 1421,
+      }
       : undefined,
     watch: {
+      // 3. tell vite to ignore watching `src-tauri`
       ignored: ["**/src-tauri/**"],
-      // (optional) if you ever see missed updates on macOS:
-      // usePolling: true,
-      // interval: 100
-    },
-    fs: {
-      allow: [
-        path.resolve(__dirname),
-        linkedPackagePath
-      ],
     },
   },
-
   resolve: {
     extensions: [".mjs", ".js", ".ts", ".jsx", ".tsx", ".json"],
-    preserveSymlinks: true,
-    dedupe: ["react", "react-dom"],
-    alias: {
-      "@bsv/brc100-ui-react-components": linkedPackagePath
-    },
+    preserveSymlinks: true
   },
-
-  optimizeDeps: {
-    include: ["@bsv/brc100-ui-react-components"],
-  },
-});
+}));
